calm-requirements.txt
pytorch>=1.6
pytorch-lightning
biopython
click
scikit-learn
numpy
requests
